requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.4
djangorestframework==3.17.1
idna==3.11
networkx==3.6.1
numpy==2.4.4
pandas==3.0.2
polyline==2.0.4
python-dateutil==2.9.0.post0
requests==2.33.1
shapely==2.1.2
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
